refactor(BridgeGame) : 게임 결과 얻기 기능 추가
+ GameResult enum클래스 패키지 도메인으로 이동
+ PlayerStatus 클래스 import문 수정
