+++ src/main/java/bridge/domain/PlayerStatus.java
@@ -1,7 +1,5 @@
 package bridge.domain;
 
-import bridge.util.GameResult;
-
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.HashMap;
